backend/requirements.txt
# backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
mongoengine==0.27.0
python-decouple==3.8
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
django-redis==5.4.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
Pillow==10.1.0
boto3==1.34.0
psutil==5.9.6
gunicorn==21.2.0
whitenoise==6.6.0